+++ CHANGES.md
@@ -9,6 +9,10 @@ This is a history of changes to [Remix](https://remix.run).
 - Treat data loader redirects the same in client-side transitions as we do in
   HTML requests.
 
+### Bug Fixes
+
+- Fixed "MaxListenersExceedWarning" warning in dev
+
 ## 0.6.2 - Fri Oct 30 2020
 
 ### Bug Fixes